+++ changelog.md
@@ -7,6 +7,7 @@
 * Fixed fishing hook not rendering for non-zombie entities
 * The pillager attack goal will no longer reset the target
 * Phantoms are no longer affected by targeting changes
+* Fixed mobs climbing always playing ladder sound
 
 ## 2.2.1
 * Creepers no longer alert their vehicle